+++ src/EnrolmentWorkflow/Request/class.RequestsGUI.php
@@ -34,12 +34,26 @@ class RequestsGUI
     const CMD_LIST_REQUESTS = "listRequests";
     const CMD_RESET_FILTER = "resetFilter";
     const GET_PARAM_REF_ID = "ref_id";
+    const GET_PARAM_REQUESTS_TYPE = "requests_type";
     const LANG_MODULE = "requests";
-    const TAB_REQUESTS = "requests";
+    const TAB_REQUESTS = "requests_";
+    const REQUESTS_TYPE_ALL = 1;
+    const REQUESTS_TYPE_OWN = 2;
+    const REQUESTS_TYPE_OPEN = 3;
+    const REQUESTS_TYPES
+        = [
+            self::REQUESTS_TYPE_ALL  => "all",
+            self::REQUESTS_TYPE_OWN  => "own",
+            self::REQUESTS_TYPE_OPEN => "open"
+        ];
     /**
      * @var int|null
      */
     protected $obj_ref_id = null;
+    /**
+     * @var int
+     */
+    protected $requests_type;
 
 
     /**
@@ -57,12 +71,14 @@ class RequestsGUI
     public function executeCommand()/*: void*/
     {
         $this->obj_ref_id = intval(filter_input(INPUT_GET, self::GET_PARAM_REF_ID));
+        $this->requests_type = intval(filter_input(INPUT_GET, self::GET_PARAM_REQUESTS_TYPE));
 
         if (!self::srUserEnrolment()->enrolmentWorkflow()->requests()->hasAccess(self::dic()->user()->getId())) {
             die("");
         }
 
         self::dic()->ctrl()->saveParameter($this, self::GET_PARAM_REF_ID);
+        self::dic()->ctrl()->saveParameter($this, self::GET_PARAM_REQUESTS_TYPE);
 
         $this->setTabs();
 
@@ -102,7 +118,9 @@ class RequestsGUI
         if (self::srUserEnrolment()->enrolmentWorkflow()->requests()->hasAccess(self::dic()->user()->getId())) {
             self::dic()->ctrl()->setParameterByClass(self::class, self::GET_PARAM_REF_ID, $obj_ref_id);
 
-            self::dic()->tabs()->addTab(self::TAB_REQUESTS, self::plugin()->translate("requests", self::LANG_MODULE), self::dic()->ctrl()
+            self::dic()->ctrl()->setParameterByClass(self::class, self::GET_PARAM_REQUESTS_TYPE, self::REQUESTS_TYPE_ALL);
+
+            self::dic()->tabs()->addTab(self::TAB_REQUESTS . self::REQUESTS_TYPE_ALL, self::plugin()->translate("requests", self::LANG_MODULE), self::dic()->ctrl()
                 ->getLinkTargetByClass([ilUIPluginRouterGUI::class, self::class], self::CMD_LIST_REQUESTS));
         }
     }
@@ -136,8 +154,11 @@ class RequestsGUI
             self::dic()->toolbar()->addButtonInstance($request_button);
         }
 
-        self::dic()->tabs()->addTab(self::TAB_REQUESTS, self::plugin()->translate("requests", self::LANG_MODULE), self::dic()->ctrl()
-            ->getLinkTargetByClass(self::class, self::CMD_LIST_REQUESTS));
+        foreach (self::REQUESTS_TYPES as $requests_type => $requests_type_lang_key) {
+            self::dic()->ctrl()->setParameterByClass(self::class, self::GET_PARAM_REQUESTS_TYPE, $requests_type);
+            self::dic()->tabs()->addTab(self::TAB_REQUESTS . $requests_type, self::plugin()->translate("type_" . $requests_type_lang_key, self::LANG_MODULE), self::dic()->ctrl()
+                ->getLinkTargetByClass(self::class, self::CMD_LIST_REQUESTS));
+        }
     }
 
 
@@ -160,7 +181,7 @@ class RequestsGUI
      */
     protected function listRequests()/*: void*/
     {
-        self::dic()->tabs()->activateTab(self::TAB_REQUESTS);
+        self::dic()->tabs()->activateTab(self::TAB_REQUESTS . $this->requests_type);
 
         $table = self::srUserEnrolment()->enrolmentWorkflow()->requests()->factory()->newTableInstance($this);
 
@@ -255,5 +276,14 @@ class RequestsGUI
     public function getObjRefId()/* : ?int*/
     {
         return $this->obj_ref_id;
+    }
+
+
+    /**
+     * @return int
+     */
+    public function getRequestsType() : int
+    {
+        return $this->requests_type;
     }
 }